GetAllOrders is returning orders from database.

+++ SweetTooth/Models/Order.cs
@@ -8,11 +8,11 @@ namespace SweetTooth.Models
     public class Order
     {
         public Guid Id { get; set; }
-        public Guid UserId { get; set; }
+        public Guid UserId { get; set; } // foreign key
         public DateTime OrderDate { get; set; }
         public int OrderNumber { get; set; }
-        public int Total { get; set; }
-        public Guid PaymentMethodId { get; set; }
+        public decimal Total { get; set; }
+        public Guid PaymentMethodId { get; set; } // foreign key
         public bool Processed { get; set; }
         public bool Shipped { get; set; }
     }
